+++ src/stores.rs
@@ -337,7 +337,7 @@ pub struct SizedCache<K, V> {
 
 impl<K, V> PartialEq for SizedCache<K, V>
 where
-    K: Eq + Hash,
+    K: Eq + Hash + Clone,
     V: PartialEq,
 {
     fn eq(&self, other: &SizedCache<K, V>) -> bool {
@@ -347,12 +347,12 @@ where
 
 impl<K, V> Eq for SizedCache<K, V>
 where
-    K: Eq + Hash,
+    K: Eq + Hash + Clone,
     V: PartialEq,
 {
 }
 
-impl<K: Hash + Eq, V> SizedCache<K, V> {
+impl<K: Hash + Eq + Clone, V> SizedCache<K, V> {
     #[deprecated(since = "0.5.1", note = "method renamed to `with_size`")]
     pub fn with_capacity(size: usize) -> SizedCache<K, V> {
         Self::with_size(size)
@@ -370,6 +370,10 @@ impl<K: Hash + Eq, V> SizedCache<K, V> {
             hits: 0,
             misses: 0,
         }
+    }
+
+    fn iter_order(&self) -> impl Iterator<Item = &(K, V)> {
+        self.order.iter()
     }
 
     /// Return an iterator of keys in the current order from most
@@ -394,37 +398,238 @@ impl<K: Hash + Eq, V> SizedCache<K, V> {
                 .expect("SizedCache::cache_set failed evicting cache key");
         }
     }
-}
 
-impl<K: Hash + Eq + Clone, V> Cached<K, V> for SizedCache<K, V> {
-    fn cache_get(&mut self, key: &K) -> Option<&V> {
+    fn get_if<F: FnOnce(&V) -> bool>(&mut self, key: &K, is_valid: F) -> Option<&V> {
         let val = self.store.get(key);
-        match val {
+        let index = match val {
+            None => None,
+            Some(index) => {
+                let val = &self.order.get(*index).1;
+                if !is_valid(val) {
+                    None
+                } else {
+                    Some(index)
+                }
+            }
+        };
+        match index {
+            None => {
+                self.misses += 1;
+                None
+            }
             Some(&index) => {
                 self.order.move_to_front(index);
                 self.hits += 1;
                 Some(&self.order.get(index).1)
             }
+        }
+    }
+
+    fn get_mut_if<F: FnOnce(&V) -> bool>(&mut self, key: &K, is_valid: F) -> Option<&mut V> {
+        let val = self.store.get(key);
+        let index = match val {
+            None => None,
+            Some(index) => {
+                let val = &self.order.get(*index).1;
+                if !is_valid(val) {
+                    None
+                } else {
+                    Some(index)
+                }
+            }
+        };
+        match index {
             None => {
                 self.misses += 1;
                 None
             }
-        }
-    }
-
-    fn cache_get_mut(&mut self, key: &K) -> std::option::Option<&mut V> {
-        let val = self.store.get(key);
-        match val {
             Some(&index) => {
                 self.order.move_to_front(index);
                 self.hits += 1;
                 Some(&mut self.order.get_mut(index).1)
             }
-            None => {
+        }
+    }
+
+    /// Get the cached value, or set it using `f` if the value
+    /// is either not-set or if `is_valid` returns `false` for
+    /// the set value.
+    ///
+    /// Returns (was_present, was_valid, mut ref to set value)
+    /// `was_valid` will be false when `was_present` is false
+    fn get_or_set_with_if<F: FnOnce() -> V, FC: FnOnce(&V) -> bool>(
+        &mut self,
+        key: K,
+        f: F,
+        is_valid: FC,
+    ) -> (bool, bool, &mut V) {
+        self.check_capacity();
+        let val = self.store.entry(key);
+        let Self { order, .. } = self;
+        match val {
+            Entry::Occupied(occupied) => {
+                let index = *occupied.get();
+                self.hits += 1;
+                let replace_existing = {
+                    let v = &order.get_mut(index).1;
+                    !is_valid(v)
+                };
+                if replace_existing {
+                    let key = occupied.key().clone();
+                    order.set(index, (key, f()));
+                }
+                order.move_to_front(index);
+                (true, !replace_existing, &mut order.get_mut(index).1)
+            }
+            Entry::Vacant(vacant) => {
+                let key = vacant.key().clone();
+                let new_val = f();
                 self.misses += 1;
-                None
+                let index = *vacant.insert(order.push_front(None));
+                order.set(index, (key, new_val));
+                (false, false, &mut order.get_mut(index).1)
             }
         }
+    }
+
+    #[allow(dead_code)]
+    fn try_get_or_set_with_if<E, F: FnOnce() -> Result<V, E>, FC: FnOnce(&V) -> bool>(
+        &mut self,
+        key: K,
+        f: F,
+        is_valid: FC,
+    ) -> Result<(bool, bool, &mut V), E> {
+        self.check_capacity();
+        let val = self.store.entry(key);
+        let Self { order, .. } = self;
+        match val {
+            Entry::Occupied(occupied) => {
+                let index = *occupied.get();
+                self.hits += 1;
+                let replace_existing = {
+                    let v = &order.get_mut(index).1;
+                    !is_valid(v)
+                };
+                if replace_existing {
+                    let key = occupied.key().clone();
+                    order.set(index, (key, f()?));
+                }
+                order.move_to_front(index);
+                Ok((true, !replace_existing, &mut order.get_mut(index).1))
+            }
+            Entry::Vacant(vacant) => {
+                let key = vacant.key().clone();
+                let new_val = f()?;
+                self.misses += 1;
+                let index = *vacant.insert(order.push_front(None));
+                order.set(index, (key, new_val));
+                Ok((false, false, &mut order.get_mut(index).1))
+            }
+        }
+    }
+}
+
+#[cfg(feature = "async")]
+impl<K, V> SizedCache<K, V>
+where
+    K: Hash + Eq + Clone + Send,
+{
+    /// Get the cached value, or set it using `f` if the value
+    /// is either not-set or if `is_valid` returns `false` for
+    /// the set value.
+    ///
+    /// Returns (was_present, was_valid, mut ref to set value)
+    /// `was_valid` will be false when `was_present` is false
+    async fn get_or_set_with_if_async<F, Fut, FC>(
+        &mut self,
+        key: K,
+        f: F,
+        is_valid: FC,
+    ) -> (bool, bool, &mut V)
+    where
+        V: Send,
+        F: FnOnce() -> Fut + Send,
+        Fut: Future<Output = V> + Send,
+        FC: FnOnce(&V) -> bool,
+    {
+        self.check_capacity();
+        let val = self.store.entry(key);
+        let Self { order, .. } = self;
+        match val {
+            Entry::Occupied(occupied) => {
+                let index = *occupied.get();
+                self.hits += 1;
+                let replace_existing = {
+                    let v = &order.get_mut(index).1;
+                    !is_valid(v)
+                };
+                if replace_existing {
+                    let key = occupied.key().clone();
+                    order.set(index, (key, f().await));
+                }
+                order.move_to_front(index);
+                (true, !replace_existing, &mut order.get_mut(index).1)
+            }
+            Entry::Vacant(vacant) => {
+                let key = vacant.key().clone();
+                let new_val = f().await;
+                self.misses += 1;
+                let index = *vacant.insert(order.push_front(None));
+                order.set(index, (key, new_val));
+                (false, false, &mut order.get_mut(index).1)
+            }
+        }
+    }
+
+    async fn try_get_or_set_with_if_async<E, F, Fut, FC>(
+        &mut self,
+        key: K,
+        f: F,
+        is_valid: FC,
+    ) -> Result<(bool, bool, &mut V), E>
+    where
+        V: Send,
+        F: FnOnce() -> Fut + Send,
+        Fut: Future<Output = Result<V, E>> + Send,
+        FC: FnOnce(&V) -> bool,
+    {
+        self.check_capacity();
+        let val = self.store.entry(key);
+        let Self { order, .. } = self;
+        match val {
+            Entry::Occupied(occupied) => {
+                let index = *occupied.get();
+                self.hits += 1;
+                let replace_existing = {
+                    let v = &order.get_mut(index).1;
+                    !is_valid(v)
+                };
+                if replace_existing {
+                    let key = occupied.key().clone();
+                    order.set(index, (key, f().await?));
+                }
+                order.move_to_front(index);
+                Ok((true, !replace_existing, &mut order.get_mut(index).1))
+            }
+            Entry::Vacant(vacant) => {
+                self.misses += 1;
+                let key = vacant.key().clone();
+                let new_val = f().await?;
+                let index = *vacant.insert(order.push_front(None));
+                order.set(index, (key, new_val));
+                Ok((false, false, &mut order.get_mut(index).1))
+            }
+        }
+    }
+}
+
+impl<K: Hash + Eq + Clone, V> Cached<K, V> for SizedCache<K, V> {
+    fn cache_get(&mut self, key: &K) -> Option<&V> {
+        self.get_if(key, |_| true)
+    }
+
+    fn cache_get_mut(&mut self, key: &K) -> std::option::Option<&mut V> {
+        self.get_mut_if(key, |_| true)
     }
 
     fn cache_set(&mut self, key: K, val: V) -> Option<V> {
@@ -437,24 +642,8 @@ impl<K: Hash + Eq + Clone, V> Cached<K, V> for SizedCache<K, V> {
     }
 
     fn cache_get_or_set_with<F: FnOnce() -> V>(&mut self, key: K, f: F) -> &mut V {
-        self.check_capacity();
-        let val = self.store.entry(key);
-        let Self { order, .. } = self;
-        match val {
-            Entry::Occupied(occupied) => {
-                let index = *occupied.get();
-                order.move_to_front(index);
-                self.hits += 1;
-                &mut order.get_mut(index).1
-            }
-            Entry::Vacant(vacant) => {
-                let key = vacant.key().clone();
-                self.misses += 1;
-                let index = *vacant.insert(order.push_front(None));
-                order.set(index, (key, f()));
-                &mut order.get_mut(index).1
-            }
-        }
+        let (_, _, v) = self.get_or_set_with_if(key, f, |_| true);
+        v
     }
 
     fn cache_remove(&mut self, k: &K) -> Option<V> {
@@ -502,24 +691,8 @@ where
         F: FnOnce() -> Fut + Send,
         Fut: Future<Output = V> + Send,
     {
-        self.check_capacity();
-        let entry = self.store.entry(k);
-        let Self { order, .. } = self;
-        match entry {
-            Entry::Occupied(occupied) => {
-                let index = *occupied.get();
-                order.move_to_front(index);
-                self.hits += 1;
-                &mut order.get_mut(index).1
-            }
-            Entry::Vacant(vacant) => {
-                let key = vacant.key().clone();
-                self.misses += 1;
-                let index = *vacant.insert(order.push_front(None));
-                order.set(index, (key, f().await));
-                &mut order.get_mut(index).1
-            }
-        }
+        let (_, _, v) = self.get_or_set_with_if_async(k, f, |_| true).await;
+        v
     }
 
     async fn try_get_or_set_with<F, Fut, E>(&mut self, k: K, f: F) -> Result<&mut V, E>
@@ -528,26 +701,7 @@ where
         F: FnOnce() -> Fut + Send,
         Fut: Future<Output = Result<V, E>> + Send,
     {
-        self.check_capacity();
-        let entry = self.store.entry(k);
-        let Self { order, .. } = self;
-        let v = match entry {
-            Entry::Occupied(occupied) => {
-                let index = *occupied.get();
-                order.move_to_front(index);
-                self.hits += 1;
-                &mut order.get_mut(index).1
-            }
-            Entry::Vacant(vacant) => {
-                self.misses += 1;
-                let value = f().await?;
-                let key = vacant.key().clone();
-                let index = *vacant.insert(order.push_front(None));
-                order.set(index, (key, value));
-                &mut order.get_mut(index).1
-            }
-        };
-
+        let (_, _, v) = self.try_get_or_set_with_if_async(k, f, |_| true).await?;
         Ok(v)
     }
 }
@@ -773,6 +927,201 @@ where
     }
 }
 
+/// Timed LRU Cache
+///
+/// Stores a limited number of values,
+/// evicting expired and least-used entries.
+/// Time expiration is determined based on entry insertion time..
+/// The TTL of an entry is not updated when retrieved.
+///
+/// Note: This cache is in-memory only
+#[derive(Clone, Debug)]
+pub struct TimedSizedCache<K, V> {
+    store: SizedCache<K, (Instant, V)>,
+    size: usize,
+    seconds: u64,
+    hits: u64,
+    misses: u64,
+}
+
+impl<K: Hash + Eq + Clone, V> TimedSizedCache<K, V> {
+    /// Creates a new `SizedCache` with a given size limit and pre-allocated backing data
+    pub fn with_size_and_lifespan(size: usize, seconds: u64) -> TimedSizedCache<K, V> {
+        if size == 0 {
+            panic!("`size` of `TimedSizedCache` must be greater than zero.")
+        }
+        TimedSizedCache {
+            store: SizedCache::with_size(size),
+            size,
+            seconds,
+            hits: 0,
+            misses: 0,
+        }
+    }
+
+    fn iter_order(&self) -> impl Iterator<Item = &(K, (Instant, V))> {
+        let max_seconds = self.seconds;
+        self.store
+            .iter_order()
+            .filter(move |(_k, stamped)| stamped.0.elapsed().as_secs() < max_seconds)
+    }
+
+    /// Return an iterator of keys in the current order from most
+    /// to least recently used.
+    /// Items passed their expiration seconds will be excluded.
+    pub fn key_order(&self) -> impl Iterator<Item = &K> {
+        self.iter_order().map(|(k, _v)| k)
+    }
+
+    /// Return an iterator of timestamped values in the current order
+    /// from most to least recently used.
+    /// Items passed their expiration seconds will be excluded.
+    pub fn value_order(&self) -> impl Iterator<Item = &(Instant, V)> {
+        self.iter_order().map(|(_k, v)| v)
+    }
+}
+
+impl<K: Hash + Eq + Clone, V> Cached<K, V> for TimedSizedCache<K, V> {
+    fn cache_get(&mut self, key: &K) -> Option<&V> {
+        let max_seconds = self.seconds;
+        let val = self
+            .store
+            .get_if(key, |stamped| stamped.0.elapsed().as_secs() < max_seconds);
+        match val {
+            None => {
+                self.misses += 1;
+                None
+            }
+            Some(stamped) => {
+                self.hits += 1;
+                Some(&stamped.1)
+            }
+        }
+    }
+
+    fn cache_get_mut(&mut self, key: &K) -> std::option::Option<&mut V> {
+        let max_seconds = self.seconds;
+        let val = self
+            .store
+            .get_mut_if(key, |stamped| stamped.0.elapsed().as_secs() < max_seconds);
+        match val {
+            None => {
+                self.misses += 1;
+                None
+            }
+            Some(stamped) => {
+                self.hits += 1;
+                Some(&mut stamped.1)
+            }
+        }
+    }
+
+    fn cache_get_or_set_with<F: FnOnce() -> V>(&mut self, key: K, f: F) -> &mut V {
+        let setter = || (Instant::now(), f());
+        let max_seconds = self.seconds;
+        let (was_present, was_valid, stamped) =
+            self.store.get_or_set_with_if(key, setter, |stamped| {
+                stamped.0.elapsed().as_secs() < max_seconds
+            });
+        if was_present && was_valid {
+            self.hits += 1;
+        } else {
+            self.misses += 1;
+        }
+        &mut stamped.1
+    }
+
+    fn cache_set(&mut self, key: K, val: V) -> Option<V> {
+        let stamped = self.store.cache_set(key, (Instant::now(), val));
+        stamped.map(|stamped| stamped.1)
+    }
+
+    fn cache_remove(&mut self, k: &K) -> Option<V> {
+        let stamped = self.store.cache_remove(k);
+        stamped.map(|stamped| stamped.1)
+    }
+    fn cache_clear(&mut self) {
+        self.store.cache_clear();
+    }
+    fn cache_reset(&mut self) {
+        self.cache_clear();
+    }
+    fn cache_size(&self) -> usize {
+        self.store.cache_size()
+    }
+    fn cache_hits(&self) -> Option<u64> {
+        Some(self.hits)
+    }
+    fn cache_misses(&self) -> Option<u64> {
+        Some(self.misses)
+    }
+    fn cache_capacity(&self) -> Option<usize> {
+        Some(self.size)
+    }
+    fn cache_lifespan(&self) -> Option<u64> {
+        Some(self.seconds)
+    }
+    fn cache_set_lifespan(&mut self, seconds: u64) -> Option<u64> {
+        let old = self.seconds;
+        self.seconds = seconds;
+        Some(old)
+    }
+}
+
+#[cfg(feature = "async")]
+#[async_trait]
+impl<K, V> CachedAsync<K, V> for TimedSizedCache<K, V>
+where
+    K: Hash + Eq + Clone + Send,
+{
+    async fn get_or_set_with<F, Fut>(&mut self, key: K, f: F) -> &mut V
+    where
+        V: Send,
+        F: FnOnce() -> Fut + Send,
+        Fut: Future<Output = V> + Send,
+    {
+        let setter = || async { (Instant::now(), f().await) };
+        let max_seconds = self.seconds;
+        let (was_present, was_valid, stamped) = self
+            .store
+            .get_or_set_with_if_async(key, setter, |stamped| {
+                stamped.0.elapsed().as_secs() < max_seconds
+            })
+            .await;
+        if was_present && was_valid {
+            self.hits += 1;
+        } else {
+            self.misses += 1;
+        }
+        &mut stamped.1
+    }
+
+    async fn try_get_or_set_with<F, Fut, E>(&mut self, key: K, f: F) -> Result<&mut V, E>
+    where
+        V: Send,
+        F: FnOnce() -> Fut + Send,
+        Fut: Future<Output = Result<V, E>> + Send,
+    {
+        let setter = || async {
+            let new_val = f().await?;
+            Ok((Instant::now(), new_val))
+        };
+        let max_seconds = self.seconds;
+        let (was_present, was_valid, stamped) = self
+            .store
+            .try_get_or_set_with_if_async(key, setter, |stamped| {
+                stamped.0.elapsed().as_secs() < max_seconds
+            })
+            .await?;
+        if was_present && was_valid {
+            self.hits += 1;
+        } else {
+            self.misses += 1;
+        }
+        Ok(&mut stamped.1)
+    }
+}
+
 impl<K: Hash + Eq, V> Cached<K, V> for HashMap<K, V> {
     fn cache_get(&mut self, k: &K) -> Option<&V> {
         self.get(k)
@@ -843,6 +1192,7 @@ mod tests {
 
     use super::SizedCache;
     use super::TimedCache;
+    use super::TimedSizedCache;
     use super::UnboundCache;
 
     #[test]
@@ -945,6 +1295,87 @@ mod tests {
     }
 
     #[test]
+    fn timed_sized_cache() {
+        let mut c = TimedSizedCache::with_size_and_lifespan(5, 2);
+        assert!(c.cache_get(&1).is_none());
+        let misses = c.cache_misses().unwrap();
+        assert_eq!(1, misses);
+
+        assert_eq!(c.cache_set(1, 100), None);
+        assert!(c.cache_get(&1).is_some());
+        let hits = c.cache_hits().unwrap();
+        let misses = c.cache_misses().unwrap();
+        assert_eq!(1, hits);
+        assert_eq!(1, misses);
+
+        assert_eq!(c.cache_set(2, 100), None);
+        assert_eq!(c.cache_set(3, 100), None);
+        assert_eq!(c.cache_set(4, 100), None);
+        assert_eq!(c.cache_set(5, 100), None);
+
+        assert_eq!(c.key_order().cloned().collect::<Vec<_>>(), [5, 4, 3, 2, 1]);
+
+        sleep(Duration::new(1, 0));
+
+        assert_eq!(c.cache_set(6, 100), None);
+        assert_eq!(c.cache_set(7, 100), None);
+
+        assert_eq!(c.key_order().cloned().collect::<Vec<_>>(), [7, 6, 5, 4, 3]);
+
+        assert!(c.cache_get(&2).is_none());
+        assert!(c.cache_get(&3).is_some());
+
+        assert_eq!(c.key_order().cloned().collect::<Vec<_>>(), [3, 7, 6, 5, 4]);
+
+        assert_eq!(2, c.cache_misses().unwrap());
+        assert_eq!(5, c.cache_size());
+
+        sleep(Duration::new(1, 0));
+
+        assert!(c.cache_get(&1).is_none());
+        assert!(c.cache_get(&2).is_none());
+        assert!(c.cache_get(&3).is_none());
+        assert!(c.cache_get(&4).is_none());
+        assert!(c.cache_get(&5).is_none());
+        assert!(c.cache_get(&6).is_some());
+        assert!(c.cache_get(&7).is_some());
+
+        assert_eq!(7, c.cache_misses().unwrap());
+
+        assert!(c.cache_set(1, 100).is_none());
+        assert!(c.cache_set(2, 100).is_none());
+        assert!(c.cache_set(3, 100).is_none());
+        assert_eq!(c.key_order().cloned().collect::<Vec<_>>(), [3, 2, 1, 7, 6]);
+
+        sleep(Duration::new(1, 0));
+
+        assert!(c.cache_get(&1).is_some());
+        assert!(c.cache_get(&2).is_some());
+        assert!(c.cache_get(&3).is_some());
+        assert!(c.cache_get(&4).is_none());
+        assert!(c.cache_get(&5).is_none());
+        assert!(c.cache_get(&6).is_none());
+        assert!(c.cache_get(&7).is_none());
+
+        assert_eq!(11, c.cache_misses().unwrap());
+
+        let mut c = TimedSizedCache::with_size_and_lifespan(5, 0);
+        let mut ticker = 0;
+        let setter = || {
+            let v = ticker;
+            ticker += 1;
+            v
+        };
+        assert_eq!(c.cache_get_or_set_with(1, setter), &0);
+        let setter = || {
+            let v = ticker;
+            ticker += 1;
+            v
+        };
+        assert_eq!(c.cache_get_or_set_with(1, setter), &1);
+    }
+
+    #[test]
     fn clear() {
         let mut c = UnboundCache::new();
 
@@ -998,6 +1429,15 @@ mod tests {
         assert_eq!(0, c.cache_size());
 
         let mut c = TimedCache::with_lifespan(3600);
+
+        assert_eq!(c.cache_set(1, 100), None);
+        assert_eq!(c.cache_set(2, 200), None);
+        assert_eq!(c.cache_set(3, 300), None);
+        c.cache_clear();
+
+        assert_eq!(0, c.cache_size());
+
+        let mut c = TimedSizedCache::with_size_and_lifespan(3, 3600);
 
         assert_eq!(c.cache_set(1, 100), None);
         assert_eq!(c.cache_set(2, 200), None);
@@ -1059,6 +1499,15 @@ mod tests {
         c.cache_reset();
 
         assert!(init_capacity <= c.store.capacity());
+
+        let mut c = TimedSizedCache::with_size_and_lifespan(init_capacity, 100);
+        assert_eq!(c.cache_set(1, 100), None);
+        assert_eq!(c.cache_set(2, 200), None);
+        assert_eq!(c.cache_set(3, 300), None);
+        assert!(init_capacity <= c.store.capacity);
+
+        c.cache_reset();
+        assert!(init_capacity <= c.store.capacity);
     }
 
     #[test]
@@ -1124,6 +1573,24 @@ mod tests {
 
         assert_eq!(Some(100), c.cache_remove(&1));
         assert_eq!(2, c.cache_size());
+
+        let mut c = TimedSizedCache::with_size_and_lifespan(3, 3600);
+
+        assert_eq!(c.cache_set(1, 100), None);
+        assert_eq!(c.cache_set(2, 200), None);
+        assert_eq!(c.cache_set(3, 300), None);
+
+        assert_eq!(Some(100), c.cache_remove(&1));
+        assert_eq!(2, c.cache_size());
+
+        assert_eq!(Some(200), c.cache_remove(&2));
+        assert_eq!(1, c.cache_size());
+
+        assert_eq!(None, c.cache_remove(&2));
+        assert_eq!(1, c.cache_size());
+
+        assert_eq!(Some(300), c.cache_remove(&3));
+        assert_eq!(0, c.cache_size());
     }
 
     #[test]
@@ -1234,13 +1701,53 @@ mod tests {
         assert_eq!(c.cache_get_or_set_with(1, || 42), &42);
 
         assert_eq!(c.cache_misses(), Some(7));
+
+        let mut c = TimedSizedCache::with_size_and_lifespan(5, 2);
+
+        assert_eq!(c.cache_get_or_set_with(0, || 0), &0);
+        assert_eq!(c.cache_get_or_set_with(1, || 1), &1);
+        assert_eq!(c.cache_get_or_set_with(2, || 2), &2);
+        assert_eq!(c.cache_get_or_set_with(3, || 3), &3);
+        assert_eq!(c.cache_get_or_set_with(4, || 4), &4);
+        assert_eq!(c.cache_get_or_set_with(5, || 5), &5);
+
+        assert_eq!(c.cache_misses(), Some(6));
+
+        assert_eq!(c.cache_get_or_set_with(0, || 0), &0);
+
+        assert_eq!(c.cache_misses(), Some(7));
+
+        assert_eq!(c.cache_get_or_set_with(0, || 42), &0);
+
+        sleep(Duration::new(1, 0));
+
+        assert_eq!(c.cache_get_or_set_with(0, || 42), &0);
+
+        assert_eq!(c.cache_get_or_set_with(1, || 1), &1);
+
+        assert_eq!(c.cache_get_or_set_with(4, || 42), &4);
+
+        assert_eq!(c.cache_get_or_set_with(5, || 42), &5);
+
+        assert_eq!(c.cache_get_or_set_with(6, || 6), &6);
+
+        assert_eq!(c.cache_misses(), Some(9));
+
+        sleep(Duration::new(1, 0));
+
+        assert_eq!(c.cache_get_or_set_with(4, || 42), &42);
+
+        assert_eq!(c.cache_get_or_set_with(5, || 42), &42);
+
+        assert_eq!(c.cache_get_or_set_with(6, || 42), &6);
+
+        assert_eq!(c.cache_misses(), Some(11));
     }
 
     #[cfg(feature = "async")]
     #[tokio::test]
     async fn test_async_trait() {
         use crate::CachedAsync;
-        assert!(1 == 1);
         let mut c = SizedCache::with_size(5);
 
         async fn _get(n: usize) -> usize {
@@ -1294,6 +1801,8 @@ mod tests {
             .try_get_or_set_with(0, || async { Ok(_try_get(10).await?) })
             .await;
         assert!(res.is_err());
+        assert!(c.key_order().next().is_none());
+
         let res: Result<&mut usize, String> = c
             .try_get_or_set_with(0, || async { Ok(_try_get(1).await?) })
             .await;
@@ -1302,5 +1811,83 @@ mod tests {
             .try_get_or_set_with(0, || async { Ok(_try_get(5).await?) })
             .await;
         assert_eq!(res.unwrap(), &1);
+    }
+
+    #[cfg(feature = "async")]
+    #[tokio::test]
+    async fn test_async_trait_timed_sized() {
+        use crate::CachedAsync;
+        let mut c = TimedSizedCache::with_size_and_lifespan(5, 1);
+
+        async fn _get(n: usize) -> usize {
+            n
+        }
+
+        assert_eq!(c.get_or_set_with(0, || async { _get(0).await }).await, &0);
+        assert_eq!(c.get_or_set_with(1, || async { _get(1).await }).await, &1);
+        assert_eq!(c.get_or_set_with(2, || async { _get(2).await }).await, &2);
+        assert_eq!(c.get_or_set_with(3, || async { _get(3).await }).await, &3);
+
+        assert_eq!(c.get_or_set_with(0, || async { _get(3).await }).await, &0);
+        assert_eq!(c.get_or_set_with(1, || async { _get(3).await }).await, &1);
+        assert_eq!(c.get_or_set_with(2, || async { _get(3).await }).await, &2);
+        assert_eq!(c.get_or_set_with(3, || async { _get(1).await }).await, &3);
+
+        sleep(Duration::new(1, 0));
+        // after sleeping, the original val should have expired
+        assert_eq!(c.get_or_set_with(0, || async { _get(3).await }).await, &3);
+
+        c.cache_reset();
+        async fn _try_get(n: usize) -> Result<usize, String> {
+            if n < 10 {
+                Ok(n)
+            } else {
+                Err("dead".to_string())
+            }
+        }
+
+        assert_eq!(
+            c.try_get_or_set_with(0, || async {
+                match _try_get(0).await {
+                    Ok(n) => Ok(n),
+                    Err(_) => Err("err".to_string()),
+                }
+            })
+            .await
+            .unwrap(),
+            &0
+        );
+        assert_eq!(
+            c.try_get_or_set_with(0, || async {
+                match _try_get(5).await {
+                    Ok(n) => Ok(n),
+                    Err(_) => Err("err".to_string()),
+                }
+            })
+            .await
+            .unwrap(),
+            &0
+        );
+
+        c.cache_reset();
+        let res: Result<&mut usize, String> = c
+            .try_get_or_set_with(0, || async { Ok(_try_get(10).await?) })
+            .await;
+        assert!(res.is_err());
+        assert!(c.key_order().next().is_none());
+
+        let res: Result<&mut usize, String> = c
+            .try_get_or_set_with(0, || async { Ok(_try_get(1).await?) })
+            .await;
+        assert_eq!(res.unwrap(), &1);
+        let res: Result<&mut usize, String> = c
+            .try_get_or_set_with(0, || async { Ok(_try_get(5).await?) })
+            .await;
+        assert_eq!(res.unwrap(), &1);
+        sleep(Duration::new(1, 0));
+        let res: Result<&mut usize, String> = c
+            .try_get_or_set_with(0, || async { Ok(_try_get(5).await?) })
+            .await;
+        assert_eq!(res.unwrap(), &5);
     }
 }